Organized to packages
Organized existing files in data to smaller sub packages.
