+++ app/src/main/java/com/waifusims/j_jlearnersdictionary/WordViewAdapter.java
@@ -1,0 +1,4 @@
+package com.waifusims.j_jlearnersdictionary;
+
+public class WordViewAdapter {
+}
