+++ app/src/main/java/data/db/RelatedWordsDbHelper.java
@@ -1,4 +1,4 @@
-package data;
+package data.db;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
