+++ app/src/main/java/data/db/VocabularyContract.java
@@ -1,4 +1,4 @@
-package data;
+package data.db;
 
 import android.provider.BaseColumns;
